untangle entities from parser code
