--- lib/ymlparser/interface.go
@@ -1,14 +1,0 @@
-package ymlparser
-
-import (
-	"context"
-
-	"github.com/santiago-labs/telophasecli/lib/awsorgs"
-)
-
-type ResourceOperation interface {
-	Call(context.Context, awsorgs.Client) error
-	ToString() string
-	AddDependent(ResourceOperation)
-	ListDependents() []ResourceOperation
-}
