+++ lib/ymlparser/organizationv2.go
@@ -6,252 +6,22 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"sort"
 	"strings"
 
-	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/organizations"
-	"github.com/aws/aws-sdk-go/service/sts"
 	"github.com/samsarahq/go/oops"
 	"github.com/santiago-labs/telophasecli/lib/awsorgs"
-	"github.com/santiago-labs/telophasecli/lib/awssess"
 	"github.com/santiago-labs/telophasecli/lib/azureorgs"
+	"github.com/santiago-labs/telophasecli/resources"
 	"gopkg.in/yaml.v3"
 )
 
 type orgDatav2 struct {
-	Organization      AccountGroup       `yaml:"Organization"`
-	AzureAccountGroup *AzureAccountGroup `yaml:"Azure,omitempty"`
+	Organization      resources.AccountGroup       `yaml:"Organization"`
+	AzureAccountGroup *resources.AzureAccountGroup `yaml:"Azure,omitempty"`
 }
 
-type AccountGroup struct {
-	ID                     *string         `yaml:"-"`
-	Name                   string          `yaml:"Name,omitempty"`
-	ChildGroups            []*AccountGroup `yaml:"AccountGroups,omitempty"`
-	Tags                   []string        `yaml:"Tags,omitempty"`
-	Accounts               []*Account      `yaml:"Accounts,omitempty"`
-	BaselineStacks         []Stack         `yaml:"Stacks,omitempty"`
-	ServiceControlPolicies []Stack         `yaml:"ServiceControlPolicies,omitempty"`
-	Parent                 *AccountGroup   `yaml:"-"`
-}
-
-type AzureAccountGroup struct {
-	// Required Fields for managing from a root subscription.
-	SubscriptionTenantID string `yaml:"SubscriptionTenantID"`
-	SubscriptionOwnerID  string `yaml:"SubscriptionOwnerID"`
-
-	// The Billing fields are combined to create a billing scope like:
-	// fmt.Sprintf("/providers/Microsoft.Billing/billingAccounts/%s/billingProfiles/%s/invoiceSections/%s",
-	// 	args.BillingAccountName,
-	// 	args.BillingProfileName,
-	// 	args.InvoiceSectionName),
-
-	// az billing account list | jq '.[] | select(.displayName == "<YOUR-BILLING-ACCOUNT-DISPLAY-NAME>") | .name'
-	BillingAccountName string `yaml:"BillingAccountName"`
-
-	// az billing profile list --account-name <billingAccountName> | jq '.[] | select(.displayName == "<YOUR-BILLING-PROFILE-DISPLAY-NAME>") | .name'
-	BillingProfileName string `yaml:"BillingProfileName"`
-
-	// az billing invoice section list --account-name <billingAccountName> --profile-name <billingProfileName> | jq '.[] | select(.displayName == "<YOUR-INVOICE-SECTION-DISPLAY-NAME>") | .name'
-	InvoiceSectionName string `yaml:"InvoiceSectionName"`
-
-	Subscriptions []Subscription `yaml:"Subscriptions,omitempty"`
-}
-
-type Subscription struct {
-	SubscriptionName string   `yaml:"Name"`
-	Account          *Account `yaml:"Account"`
-
-	BaselineStacks []Stack `yaml:"Stacks,omitempty"`
-}
-
-func (grp AccountGroup) AllTags() []string {
-	var tags []string
-	tags = append(tags, grp.Tags...)
-	if grp.Parent != nil {
-		tags = append(tags, grp.Parent.AllTags()...)
-	}
-	return tags
-}
-
-func (grp AccountGroup) AllBaselineStacks() []Stack {
-	var stacks []Stack
-	if grp.Parent != nil {
-		stacks = append(stacks, grp.Parent.AllBaselineStacks()...)
-	}
-	stacks = append(stacks, grp.BaselineStacks...)
-	return stacks
-}
-
-// grp == configuration in organization.yml.
-func (grp AccountGroup) Diff(orgClient awsorgs.Client) []ResourceOperation {
-	// Order of operations matters. Groups must be created first, followed by account creation,
-	// and finally (re)parenting groups and accounts.
-	var operations []ResourceOperation
-
-	stsClient := sts.New(session.Must(awssess.DefaultSession()))
-	caller, err := stsClient.GetCallerIdentity(&sts.GetCallerIdentityInput{})
-	if err != nil {
-		panic(err)
-	}
-
-	providerRootGroup, err := FetchGroupAndDescendents(context.TODO(), orgClient, *grp.ID, *caller.Account)
-	if err != nil {
-		panic(err)
-	}
-
-	providerGroups := providerRootGroup.AllDescendentGroups()
-	for _, parsedGroup := range grp.AllDescendentGroups() {
-		var found bool
-		for _, providerGroup := range providerGroups {
-			if parsedGroup.ID != nil && *providerGroup.ID == *parsedGroup.ID {
-				found = true
-				if parsedGroup.Parent.ID == nil {
-					for _, newGroup := range FlattenOperations(operations) {
-						newGroupOperation, ok := newGroup.(*OrganizationUnitOperation)
-						if !ok {
-							continue
-						}
-						if newGroupOperation.OrganizationUnit == parsedGroup.Parent {
-							newGroup.AddDependent(&OrganizationUnitOperation{
-								OrganizationUnit: parsedGroup,
-								NewParent:        parsedGroup.Parent,
-								CurrentParent:    providerGroup.Parent,
-								Operation:        UpdateParent,
-							})
-						}
-					}
-
-				} else if *parsedGroup.Parent.ID != *providerGroup.Parent.ID {
-					operations = append(operations, &OrganizationUnitOperation{
-						OrganizationUnit: parsedGroup,
-						NewParent:        parsedGroup.Parent,
-						CurrentParent:    providerGroup.Parent,
-						Operation:        UpdateParent,
-					})
-				}
-				break
-			}
-		}
-
-		if !found {
-			if parsedGroup.Parent.ID == nil {
-				for _, newGroup := range FlattenOperations(operations) {
-					newGroupOperation, ok := newGroup.(*OrganizationUnitOperation)
-					if !ok {
-						continue
-					}
-					if newGroupOperation.OrganizationUnit == parsedGroup.Parent {
-						newGroup.AddDependent(&OrganizationUnitOperation{
-							OrganizationUnit: parsedGroup,
-							NewParent:        parsedGroup.Parent,
-							Operation:        Create,
-						})
-					}
-				}
-			} else {
-				operations = append(operations, &OrganizationUnitOperation{
-					OrganizationUnit: parsedGroup,
-					NewParent:        parsedGroup.Parent,
-					Operation:        Create,
-				})
-			}
-		}
-	}
-
-	providerAccounts := providerRootGroup.AllDescendentAccounts()
-	for _, parsedAcct := range grp.AllDescendentAccounts() {
-		var found bool
-		for _, providerAcct := range providerAccounts {
-			if providerAcct.Email == parsedAcct.Email {
-				found = true
-				if parsedAcct.Parent.ID == nil {
-					for _, newGroup := range FlattenOperations(operations) {
-						newGroupOperation, ok := newGroup.(*OrganizationUnitOperation)
-						if !ok {
-							continue
-						}
-						if newGroupOperation.OrganizationUnit == parsedAcct.Parent {
-							newGroup.AddDependent(&AccountOperation{
-								Account:       parsedAcct,
-								Operation:     UpdateParent,
-								CurrentParent: providerAcct.Parent,
-								NewParent:     parsedAcct.Parent,
-							})
-						}
-					}
-				} else if *providerAcct.Parent.ID != *parsedAcct.Parent.ID {
-					operations = append(operations, &AccountOperation{
-						Account:       parsedAcct,
-						NewParent:     parsedAcct.Parent,
-						CurrentParent: providerAcct.Parent,
-						Operation:     UpdateParent,
-					})
-				}
-				break
-			}
-		}
-
-		if !found {
-			if parsedAcct.Parent.ID == nil {
-				for _, newGroup := range FlattenOperations(operations) {
-					newGroupOperation, ok := newGroup.(*OrganizationUnitOperation)
-					if !ok {
-						continue
-					}
-					if newGroupOperation.OrganizationUnit == parsedAcct.Parent {
-						newGroup.AddDependent(&AccountOperation{
-							Account:   parsedAcct,
-							Operation: Create,
-							NewParent: parsedAcct.Parent,
-						})
-					}
-				}
-			} else {
-				operations = append(operations, &AccountOperation{
-					Account:   parsedAcct,
-					Operation: Create,
-					NewParent: parsedAcct.Parent,
-				})
-			}
-		}
-	}
-
-	return operations
-}
-
-func (grp AccountGroup) AllDescendentAccounts() []*Account {
-	var accounts []*Account
-	accounts = append(accounts, grp.Accounts...)
-
-	for _, group := range grp.ChildGroups {
-		accounts = append(accounts, group.AllDescendentAccounts()...)
-	}
-
-	sort.Slice(accounts, func(i, j int) bool {
-		return accounts[i].Email < accounts[j].Email
-	})
-
-	return accounts
-}
-
-func (grp AccountGroup) AllDescendentGroups() []*AccountGroup {
-	var groups []*AccountGroup
-	groups = append(groups, grp.ChildGroups...)
-
-	for _, group := range grp.ChildGroups {
-		groups = append(groups, group.AllDescendentGroups()...)
-	}
-
-	sort.Slice(groups, func(i, j int) bool {
-		return groups[i].Name < groups[j].Name
-	})
-
-	return groups
-
-}
-
-func ParseOrganizationV2(filepath string) (*AccountGroup, *AzureAccountGroup, error) {
+func ParseOrganizationV2(filepath string) (*resources.AccountGroup, *resources.AzureAccountGroup, error) {
 	if filepath == "" {
 		return nil, nil, errors.New("filepath is empty")
 	}
@@ -306,7 +76,7 @@ func ParseOrganizationV2(filepath string) (*AccountGroup, *AzureAccountGroup, er
 	return &org.Organization, azureGroup, nil
 }
 
-func hydrateSubscriptions(subsClient *azureorgs.Client, azureGroup *AzureAccountGroup) error {
+func hydrateSubscriptions(subsClient *azureorgs.Client, azureGroup *resources.AzureAccountGroup) error {
 	subs, err := subsClient.CurrentSubscriptions(context.TODO())
 	if err != nil {
 		return oops.Wrapf(err, "")
@@ -315,10 +85,10 @@ func hydrateSubscriptions(subsClient *azureorgs.Client, azureGroup *AzureAccount
 	for i, sub := range azureGroup.Subscriptions {
 		for _, liveSub := range subs {
 			if sub.SubscriptionName == *liveSub.DisplayName {
-				azureGroup.Subscriptions[i].Account = &Account{
+				azureGroup.Subscriptions[i].Account = &resources.Account{
 					SubscriptionID: strings.Split(*liveSub.ID, "/")[2],
 					AccountName:    *liveSub.DisplayName,
-					BaselineStacks: sub.BaselineStacks,
+					BaselineStacks: sub.Account.BaselineStacks,
 				}
 			}
 		}
@@ -327,7 +97,7 @@ func hydrateSubscriptions(subsClient *azureorgs.Client, azureGroup *AzureAccount
 	return nil
 }
 
-func hydrateAccount(group *AccountGroup, acct *organizations.Account) {
+func hydrateAccount(group *resources.AccountGroup, acct *organizations.Account) {
 	for idx, parsedAcct := range group.Accounts {
 		group.Accounts[idx].Parent = group
 		if parsedAcct.Email == *acct.Email {
@@ -341,7 +111,7 @@ func hydrateAccount(group *AccountGroup, acct *organizations.Account) {
 	}
 }
 
-func hydrateOU(orgClient awsorgs.Client, group *AccountGroup, ou *organizations.OrganizationalUnit) error {
+func hydrateOU(orgClient awsorgs.Client, group *resources.AccountGroup, ou *organizations.OrganizationalUnit) error {
 	if ou != nil {
 		group.ID = ou.Id
 		children, err := orgClient.GetOrganizationUnitChildren(context.TODO(), *group.ID)
@@ -383,65 +153,7 @@ func hydrateOU(orgClient awsorgs.Client, group *AccountGroup, ou *organizations.
 	return nil
 }
 
-func FetchGroupAndDescendents(ctx context.Context, orgClient awsorgs.Client, ouID, mgmtAccountID string) (AccountGroup, error) {
-	var group AccountGroup
-
-	var providerGroup *organizations.OrganizationalUnit
-
-	// we treat the root group as an OU, but AWS does not consider root as an OU.
-	if strings.HasPrefix(ouID, "r-") {
-		name := "root"
-		providerGroup = &organizations.OrganizationalUnit{
-			Id:   &ouID,
-			Name: &name,
-		}
-	} else {
-		var err error
-		providerGroup, err = orgClient.GetOrganizationUnit(ctx, ouID)
-		if err != nil {
-			return group, err
-		}
-	}
-
-	group.ID = &ouID
-	group.Name = *providerGroup.Name
-
-	groupAccounts, err := orgClient.CurrentAccountsForParent(ctx, *group.ID)
-	if err != nil {
-		return group, err
-	}
-
-	for _, providerAcct := range groupAccounts {
-		acct := Account{
-			AccountID:   *providerAcct.Id,
-			Email:       *providerAcct.Email,
-			Parent:      &group,
-			AccountName: *providerAcct.Name,
-		}
-		if providerAcct.Id == &mgmtAccountID {
-			acct.ManagementAccount = true
-		}
-		group.Accounts = append(group.Accounts, &acct)
-	}
-
-	children, err := orgClient.GetOrganizationUnitChildren(ctx, ouID)
-	if err != nil {
-		return group, err
-	}
-
-	for _, providerChild := range children {
-		child, err := FetchGroupAndDescendents(ctx, orgClient, *providerChild.Id, mgmtAccountID)
-		if err != nil {
-			return group, err
-		}
-		child.Parent = &group
-		group.ChildGroups = append(group.ChildGroups, &child)
-	}
-
-	return group, nil
-}
-
-func WriteOrgV2File(filepath string, org *AccountGroup) error {
+func WriteOrgV2File(filepath string, org *resources.AccountGroup) error {
 	orgData := orgDatav2{
 		Organization: *org,
 	}
@@ -461,7 +173,7 @@ func WriteOrgV2File(filepath string, org *AccountGroup) error {
 	return nil
 }
 
-func validOrganizationV2(data AccountGroup) error {
+func validOrganizationV2(data resources.AccountGroup) error {
 	accountEmails := map[string]struct{}{}
 
 	validStates := []string{"delete", ""}
@@ -482,65 +194,6 @@ func validOrganizationV2(data AccountGroup) error {
 	}
 
 	return nil
-}
-
-func FlattenOperations(topList []ResourceOperation) []ResourceOperation {
-	var finalOperations []ResourceOperation
-
-	for _, op := range topList {
-		finalOperations = append(finalOperations, op)
-		finalOperations = append(finalOperations, FlattenOperations(op.ListDependents())...)
-	}
-
-	return finalOperations
-}
-
-func (az AzureAccountGroup) Diff(subscriptionClient *azureorgs.Client) ([]ResourceOperation, error) {
-	if subscriptionClient == nil {
-		return nil, nil
-	}
-
-	ctx := context.TODO()
-	subscriptions, err := subscriptionClient.CurrentSubscriptions(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var operations []ResourceOperation
-
-	liveSubs := map[string]struct{}{}
-	for _, sub := range subscriptions {
-		liveSubs[*sub.DisplayName] = struct{}{}
-	}
-
-	subsToCreate := map[string]Subscription{}
-	for _, iacSub := range az.Subscriptions {
-		if _, ok := liveSubs[iacSub.SubscriptionName]; !ok {
-			subsToCreate[iacSub.SubscriptionName] = iacSub
-		}
-	}
-
-	for i := range subsToCreate {
-		toCreate := subsToCreate[i]
-
-		operations = append(operations, &AzureSubscriptionOperation{
-			Operation:    Create,
-			Subscription: &toCreate,
-			AzureGroup:   az,
-		})
-	}
-
-	return operations, nil
-}
-
-func (az AzureAccountGroup) AllDescendentAccounts() []*Account {
-	var accounts []*Account
-
-	for i := range az.Subscriptions {
-		accounts = append(accounts, az.Subscriptions[i].Account)
-	}
-
-	return accounts
 }
 
 func isOneOf(s string, valid ...string) bool {
