+++ resources/subscription.go
@@ -1,0 +1,6 @@
+package resources
+
+type Subscription struct {
+	SubscriptionName string   `yaml:"Name"`
+	Account          *Account `yaml:"Account"`
+}
